+++ src/cli/docs/docs.test.ts
@@ -2,8 +2,8 @@ import { describe, expect, it } from "@jest/globals";
 import path from "path";
 import * as childProcess from "child_process";
 import { promisify } from "util";
-import { getDocsForCommand } from "./docs.markdown.helper";
-import { parseCLIOutput } from "./docs.cli.helper";
+import { getDocsForCommand } from "./markdown.helper";
+import { parseCLIOutput } from "./cli.helper";
 
 // Can probably simplify this. Tests take 30ish secs to run right now
 const execPromise = promisify(childProcess.exec);
